+++ app/(tabs)/folders/SearchSection.tsx
