+++ app/(tabs)/folders/_components/folderCard/FolderCard.tsx
