+++ app/(tabs)/folders/_components/folderCard/ScrapButton.tsx
