+++ app/(tabs)/folders/_components/restaurantCard/RestaurantCard.stories.tsx
