+++ app/(tabs)/folders/_components/restaurantCard/RestaurantCard.tsx
